cms-alfresco-platform-jar/src/main/java/gov/uspto/trademark/cms/repo/services/impl/cabinet/cmscase/base/doctype/TeasPdfDocService.java
package gov.uspto.trademark.cms.repo.services.impl.cabinet.cmscase.base.doctype;

import java.io.Serializable;
import java.util.Map;

import org.alfresco.service.cmr.repository.ContentReader;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import gov.uspto.trademark.cms.repo.constants.TradeMarkModel;
import gov.uspto.trademark.cms.repo.filters.CmsDataFilter;
import gov.uspto.trademark.cms.repo.helpers.TeasPdfNodeCreator;
import gov.uspto.trademark.cms.repo.model.cabinet.cmscase.TeasPdf;
import gov.uspto.trademark.cms.repo.nodelocator.TeasPdfNodeLocator;
import gov.uspto.trademark.cms.repo.services.impl.cabinet.cmscase.AbstractCaseCommonService;
import gov.uspto.trademark.cms.repo.services.util.ContentItem;
import gov.uspto.trademark.cms.repo.webscripts.beans.TmngAlfResponse;

/**
 * This service will implement the teaspdf document type functionality.
 * 
 * @author stank
 *
 */
@Component("TeasPdfDocService")
public class TeasPdfDocService extends AbstractCaseCommonService {

    /**
     * Instantiates a new teaspdf doc service.
     *
     * @param caseNodeCreator
     *            the case node creator
     * @param caseNodeLocator
     *            the case node locator
     */
    @Autowired
    public TeasPdfDocService(TeasPdfNodeCreator teasPdfNodeCreator, TeasPdfNodeLocator teasPdfNodeLocator) {
        super(teasPdfNodeCreator, teasPdfNodeLocator);
    }

    /*
     * (non-Javadoc)
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#create(java.lang
     * .String, java.lang.String, byte[], java.util.Map)
     */
    @Override
    public TmngAlfResponse create(final String id, final String fileName, final ContentItem content,
            Map<String, Serializable> properties) {
        return this.createDocument(id, fileName, content, TradeMarkModel.TEAS_PDF_QNAME, properties, TeasPdf.class);
    }

    /*
     * (non-Javadoc)
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#update(java.lang
     * .String, java.lang.String,
     * gov.uspto.trademark.cms.repo.services.util.ContentItem, java.util.Map)
     */
    @Override
    public TmngAlfResponse update(String id, String fileName, ContentItem content, Map<String, Serializable> properties) {
        return updateDocument(id, fileName, content, TradeMarkModel.TEAS_PDF_QNAME, properties, TeasPdf.class);
    }

    /*
     * (non-Javadoc)
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#updateMetadata
     * (java.lang.String, java.lang.String, java.util.Map)
     */
    @Override
    public TmngAlfResponse updateMetadata(String id, String fileName, Map<String, Serializable> properties) {
        return null;
    }

    /*
     * (non-Javadoc)
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#retrieveContent
     * (java.lang.String, java.lang.String, java.lang.String)
     */
    @Override
    public ContentReader retrieveContent(String id, String fileName, String versionNumber) {
        return retrieveContent(cmsNodeLocator, TradeMarkModel.TEAS_PDF_QNAME, id, fileName, versionNumber);
    }

    /*
     * (non-Javadoc)
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#retrieveContent
     * (java.lang.String, java.lang.String, java.lang.String,
     * gov.uspto.trademark.cms.repo.filters.CmsDataFilter)
     */
    @Override
    public ContentReader retrieveContent(String id, String fileName, String versionNumber, CmsDataFilter dataFilter) {
        return retrieveContent(cmsNodeLocator, TradeMarkModel.TEAS_PDF_QNAME, id, fileName, versionNumber, dataFilter);
    }

    /*
     * This method is used to retrieve the metadata of the registration
     * certificate for eog
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#retrieveMetadata
     * (java.lang.String, java.lang.String, java.lang.String)
     */
    @Override
    public byte[] retrieveMetadata(String id, String fileName, String versionNumber) {
        return retrieveMetadata(cmsNodeLocator, id, fileName, versionNumber, TradeMarkModel.TEAS_PDF_QNAME, TeasPdf.class);
    }

    /*
     * This method is used to retrieve the metadata of the registration
     * certificate for eog
     * 
     * @see
     * gov.uspto.trademark.cms.repo.services.CmsCommonService#retrieveMetadata
     * (java.lang.String, java.lang.String, java.lang.String,
     * gov.uspto.trademark.cms.repo.filters.CmsDataFilter)
     */
    @Override
    public byte[] retrieveMetadata(String id, String fileName, String versionNumber, CmsDataFilter dataFilter) {
        return retrieveMetadata(cmsNodeLocator, id, fileName, versionNumber, TradeMarkModel.TEAS_PDF_QNAME, TeasPdf.class);
    }

    /**
     * @Title: retrieveAttachmentContent @Description: @param id @param
     * fileName @param versionNumber @return @return ContentReader @throws
     */
    public ContentReader retrieveAttachmentContent(String id, String fileName, String versionNumber) {
        return retrieveContent(cmsNodeLocator, TradeMarkModel.TEAS_PDF_QNAME, id, fileName, versionNumber);
    }

}
